Working for FIFO. Bottleneck in process search. Attempting to fix.

+++ statistics.h
@@ -2,7 +2,7 @@
 #define STATS_H
 
 typedef struct stats_t {
-	unsigned long int sum_page_frames;
+	double sum_page_frames;
 	double average_memory_utilization;
 	unsigned long int sum_runnable_processes;
 	double average_runnable_processes;
